Upgrade from Tauri v1 to v2
- Update Cargo.toml dependencies to Tauri v2.5.1 with latest plugin versions
- Migrate tauri.conf.json to v2 structure (rename 'tauri' to 'app', restructure bundle config)
- Add mobile support by converting main.rs to lib.rs with mobile_entry_point macro
- Update frontend dependencies to use plugin-based APIs
- Replace @tauri-apps/api imports with plugin-specific imports
- Update SQL plugin from v1 git branch to v2.2.1 stable release
- Replace writeBinaryFile with writeFile API in PDF export functionality
