+++ package.json
@@ -15,6 +15,9 @@
     "@lingui/macro": "^4.8.0",
     "@lingui/react": "^4.8.0",
     "@react-pdf/renderer": "^3.4.2",
+    "@tauri-apps/plugin-dialog": "^2.2.2",
+    "@tauri-apps/plugin-fs": "^2.3.0",
+    "@tauri-apps/plugin-sql": "^2.2.1",
     "antd": "^5.19.0",
     "axios": "^1.7.2",
     "jotai": "^2.8.4",
@@ -26,13 +29,12 @@
     "react-dom": "^18",
     "react-pdf": "^8",
     "react-router-dom": "^6",
-    "tauri-plugin-sql-api": "https://github.com/tauri-apps/tauri-plugin-sql#v1",
     "usehooks-ts": "^3.1.0"
   },
   "devDependencies": {
     "@lingui/cli": "^4.8.0",
     "@lingui/vite-plugin": "^4.8.0",
-    "@tauri-apps/cli": "^1.5.11",
+    "@tauri-apps/cli": "^2.0.0",
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18.2.21",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
