+++ src-tauri/Cargo.toml
@@ -11,22 +11,25 @@ rust-version = "1.75"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+name = "upcount_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 [build-dependencies]
-tauri-build = { version = "1.5.3", features = [] }
+tauri-build = { version = "2.2.0", features = [] }
 
 [dependencies]
 include_dir = { version = "0.7.4", features = ["glob"] }
 serde_json = "1.0"
 serde = { version = "1.0.203", features = ["derive"] }
-tauri = { version = "1.7.1", features = [ "updater", "fs-all", "dialog-all"] }
+tauri = { version = "2.5.1", features = ["rustls-tls"] }
+tauri-plugin-updater = "2.8.1"
+tauri-plugin-dialog = "2.2.2"
+tauri-plugin-fs = "2.3.0"
+tauri-plugin-sql = { version = "2.2.1", features = ["sqlite"] }
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
 # If you use cargo directly instead of tauri's cli you can use this feature flag to switch between tauri's `dev` and `build` modes.
 # DO NOT REMOVE!!
 custom-protocol = [ "tauri/custom-protocol" ]
-
-[dependencies.tauri-plugin-sql]
-git = "https://github.com/tauri-apps/plugins-workspace"
-branch = "v1"
-features = ["sqlite"]
